Improved chaincore/magic-block package
